Migrate from LESS to SCSS
One of hidden differences between LESS and SCSS is that in LESS mixins are called lazily, and in SCSS -- eagerly.
Because of this, the migration required a change in the `define-style` mixin.
Now, instead of calling `state-default` and `state-hover` mixins defined by the calling code before each call to the `define-style` mixin,
the `define-style` mixin repeatedly invokes its content block, each time with a different set of arguments passed through global variables.
This content block is passed by the calling code like this: `@include define-style(...) { /* content block */ }`.